Reformat basic traits, add ENeg operations
