+++ models/src/main/scala/coop/rchain/models/rholangN/Operation.scala
@@ -1,0 +1,4 @@
+package coop.rchain.models.rholangN
+
+final class ENegN(val p: ParN) extends OperationN
+object ENegN { def apply(p: ParN): ENegN = new ENegN(p) }
