+++ models/src/test/scala/coop/rchain/models/rholangN/ParSpec.scala
@@ -38,13 +38,14 @@ class ParSpec extends AnyFlatSpec with ScalaCheckPropertyChecks with Matchers {
   val bytesTest: Array[Byte] = Array.fill(sizeTest)(42)
   val strTest: String        = List.fill(sizeTest)("42").mkString
 
-  /** Main types */
+  /** Par */
   it should "test ParProc" in {
     val p1 = ParProcN(Seq(GNilN(), ParProcN()))
     val p2 = ParProcN(Seq(ParProcN(), GNilN()))
     simpleCheck(p1, Some(p2)) should be(true)
   }
 
+  /** Basic types */
   it should "test Send with same data order" in {
     val p = SendN(GNilN(), Seq(GNilN(), SendN(GNilN(), GNilN())), persistent = true)
     simpleCheck(p) should be(true)
@@ -173,7 +174,12 @@ class ParSpec extends AnyFlatSpec with ScalaCheckPropertyChecks with Matchers {
     simpleCheck(p) should be(true)
   }
 
-  /** Expr */
+  /** Operations */
+  it should "test ENeg" in {
+    val p = ENegN(GIntN(42))
+    simpleCheck(p) should be(true)
+  }
+
   /** Bundle */
   /** Connective */
   /** Other types */
